Create basic resolver without DTOs
